requirements.txt
argparse
requests>=2.22.0
cmd2>=0.9.18
protobuf>=3.10.0
packaging>=19.2
oauth2client>=4.1.3
click>=5.0
pyserial>=3.0
future>=0.15.2
cryptography>=2.1.4
pyparsing>=2.0.3,<2.4.0
pyelftools>=0.22